Move more helpers out of JSX files; add no-undef rule and ignore test files
