+++ editor/src/views/projects/projectView/ProjectComponents/MoveComponentForm.jsx
@@ -1,4 +1,3 @@
-import React, { useMemo } from "react";
 import { useQuery } from "@apollo/client";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
@@ -9,23 +8,6 @@ import { getOptionLabel, isOptionEqualToValue } from "./utils/form";
 import { PROJECT_OPTIONS } from "src/queries/project";
 
 import * as yup from "yup";
-
-/**
- * Take the moped_project records data response and create options for a MUI autocomplete
- * @param {Object} data Data returned with moped_project records
- * @returns {Array} The options with value and label for moped projects
- */
-export const useProjectOptions = (data) =>
-  useMemo(() => {
-    if (!data) return [];
-
-    const options = data.moped_project.map((option) => ({
-      value: option.project_id,
-      label: `${option.project_id} - ${option.project_name_full}`,
-    }));
-
-    return options;
-  }, [data]);
 
 const validationSchema = yup.object().shape({
   project: yup.object().required(),
@@ -49,7 +31,7 @@ const MoveComponentForm = ({ onSave, component }) => {
   });
   const projectOptions = useProjectOptions(data);
 
-  error && console.error(error);
+  if (error) console.error(error);
 
   return (
     <form onSubmit={handleSubmit(onSave)}>
